requirements.txt
# ── emotion-detection-action SDK ─────────────────────────────────────────────
# Installed directly from GitHub. Pin to a commit SHA in production.
emotion-detection-action @ git+https://github.com/saurabh947/emotion-detection-action.git

# ── gRPC / Protobuf ───────────────────────────────────────────────────────────
grpcio==1.62.1
grpcio-tools==1.62.1        # protoc compiler (used at build time via `make proto`)
protobuf==4.25.3

# ── TorchServe ────────────────────────────────────────────────────────────────
# Base image is pytorch/pytorch (not pytorch/torchserve), so install explicitly.
torchserve==0.12.0
torch-model-archiver==0.12.0
nvgpu==0.10.0                   # TorchServe GPU metrics collector dependency

# ── HTTP client (gRPC server → TorchServe loopback) ──────────────────────────
httpx==0.27.0

# ── Config ────────────────────────────────────────────────────────────────────
pydantic==2.6.4
pydantic-settings==2.2.1

# ── GCS ───────────────────────────────────────────────────────────────────────
google-cloud-storage==2.16.0

# ── Numerics (inherited from SDK, pinned for reproducibility) ─────────────────
numpy==1.26.4
